Removed Useless Class and added a not working mixin.
