+++ common/src/main/resources/valkyrienskies-common.mixins.json
@@ -16,6 +16,7 @@
     "feature.container_distance_check.MixinRandomizableContainerBlockEntity",
     "feature.dispensers.MixinDefaultDispenseItemBehavior",
     "feature.distance_replace.MixinEntity",
+    "feature.entity_collision.MixinEntity",
     "feature.explosions.MixinExplosion",
     "feature.fire_between_ship_and_world.LavaFluidMixin",
     "feature.fluid_escaping_ship_config.MixinFlowingFluid",
